Move rollover tests into mod for reusability
